requirements.txt
python-telegram-bot>=20
requests
python-dotenv
